+++ velox/experimental/stateful/state/KeyedStateBackend.h
@@ -49,8 +49,8 @@ class KeyedStateBackend : public Snapshotable, public CheckpointListener {
   getOrCreateRankMapState(StateDescriptor& stateDescriptor) = 0;
 
   // TODO: Flink create PriorityQueue.
-  virtual std::shared_ptr<InternalTimerService<uint32_t, int64_t>>
-  createTimerService(Triggerable<uint32_t, int64_t>* triggerable) = 0;
+  virtual std::shared_ptr<InternalTimerService<int64_t, int64_t>>
+  createTimerService(Triggerable<int64_t, int64_t>* triggerable) = 0;
 
   virtual std::shared_ptr<InternalTimerService<uint32_t, TimeWindow>>
   createGroupWindowAggTimerService(
